fix: Fix broken upload in web client
The existing code used File instances to handle image uploads. This, however, does not work in web browsers since files uploaded to a browser do not expose their full path to the browser.

This commit switches the implementation from File to NamedByteArray. NamedByteArray is a tuple of a filename and a byte array. In other words, a read-only file that does not have a directory.