Remove Bearer token on logout
After calling `DELETE /api/auth` the auth token should be considered invalid.

Trying to re-use it for login will fail when the logout succeeded server
side.

If the logout did not succeed (e.g. 401 Unauthorized), it is safe to
assume the token is no longer valid anyway.

Also, properly update the connection state (which disconnects from the socket).

Fixes #855